move all create sound list function to sound.ts
